requirements-nano.txt
absl-py==0.12.0
aiocontextvars==0.2.2
astor==0.8.1
astunparse==1.6.3
attrs==21.2.0
beautifulsoup4==4.10.0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
contextvars==2.4
Cython==0.29.25
cython-bbox==0.1.3
dataclasses==0.8
deepface==0.0.68
filelock==3.4.0
Flask==2.0.2
flatbuffers==1.12
future==0.18.2
gast==0.4.0
gdown==4.2.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.34.1
h5py==2.10.0
idna==3.3
immutables==0.16
importlib-metadata==4.8.2
itsdangerous==2.0.1
Jinja2==3.0.3
keras==2.7.0
Keras-Applications==1.0.8
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
lap==0.4.0
loguru==0.5.3
Markdown==3.3.6
MarkupSafe==2.0.1
mediapipe @ file:///home/nvidia/Dreamsoft/ThwackRetail/mediapipe-bin/mediapipe-0.8.5_cuda102-cp36-none-linux_aarch64.whl
mtcnn==0.1.1
numpy==1.19.4
nvidia-pyindex==1.0.9
oauthlib==3.1.1
onnxruntime-gpu @ file:///home/nvidia/Downloads/onnxruntime_gpu-1.7.0-cp36-cp36m-linux_aarch64.whl
opencv-contrib-python==4.5.4.60
opencv-python==4.5.4.60
opt-einsum==3.3.0
pandas==1.1.5
Pillow==8.4.0
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2018.3
requests==2.26.0
requests-oauthlib==1.3.0
retina-face==0.0.5
rsa==4.8
scipy==1.5.4
six==1.15.0
soupsieve==2.3.1
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0+nv21.8
tensorflow-estimator==2.5.0
tensorrt==0.0.1
termcolor==1.1.0
thop==0.0.31.post2005241907
torch @ file:///home/nvidia/Dreamsoft/ThwackRetail/torch-1.10.0-cp36-cp36m-linux_aarch64.whl
torchvision==0.11.1
tqdm==4.62.3
typing-extensions==3.7.4.3
urllib3==1.26.7
Werkzeug==2.0.2
wrapt==1.12.1
zipp==3.6.0
